+++ speeduino/display.ino
@@ -187,8 +187,8 @@ void updateDisplay()
       break;
   }
 
-  int barWidth = ldiv(((unsigned long)currentStatus.RPM * 128), 9000).quot;
-  //int barWidth = map(currentStatus.RPM, 0, 9000, 0, 128);
+  int16_t barWidth = ldiv(((unsigned long)currentStatus.RPM * 128), 9000).quot;
+  //int16_t barWidth = map(currentStatus.RPM, 0, 9000, 0, 128);
   display.fillRect(0, 20, barWidth, 10, 1);
 
   display.display();
